+++ CHANGELOG.md
@@ -26,6 +26,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 -   Fixed inactive Grid slots still rendering resources.
 -   Fixed being able to interact with inactive Grid.
 -   Fixed nearly on/off Controller model not being rendered correctly on Forge.
+-   Fixed Controller energy tooltip not working.
 
 ## [2.0.0-milestone.2.14] - 2023-08-19